Fix issue #9306: Move "Open only one instance of JabRef" preference option to somewhere else
